prepare for migration to new runtime metrics
