+++ instrumentation/runtime/go.mod
@@ -3,12 +3,16 @@ module go.opentelemetry.io/contrib/instrumentation/runtime
 go 1.21
 
 require (
+	github.com/stretchr/testify v1.9.0
 	go.opentelemetry.io/otel v1.27.0
 	go.opentelemetry.io/otel/metric v1.27.0
 )
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	go.opentelemetry.io/otel/trace v1.27.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
